+++ src/server/index.js
@@ -2,7 +2,6 @@
 
 require('dotenv').config();
 
-const logout = require('./endpoints/logout-endpoints');
 const signUp = require('./endpoints/signup-endpoints');
 const bodyParser = require('body-parser');
 const DatabaseHealth = require('./endpoints/heartbeat');
@@ -17,7 +16,7 @@ let portNum = process.env.PORT || defaultPortNum;
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(express.static('dist'));
-app.delete('/api/logout', logout.logout);
+app.delete('/api/login', loginController.logout);
 app.post('/api/signup', signUp.signupUser);
 app.post('/api/login', jsonParser, loginController.login);
 app.get('/heartbeat', DatabaseHealth.checkDatabaseHealth);
